// src/util/types.ts
export enum KlerosItemStatus {
  Absent = 'Absent',
  Registered = 'Registered',
  RegistrationRequested = 'RegistrationRequested',
  ClearingRequested = 'ClearingRequested',
}
// ...
export interface BalanceItem {
  outcomeName: string
  probability: number
  currentPrice: number
  currentDisplayPrice?: number
  shares: BigNumber
  payout: Big
  holdings: BigNumber
}

export interface BondItem {
  outcomeName: string
  bondedEth: BigNumber
}

export interface AnswerItem {
  answer: string
  bondAggregate: BigNumber
}

export interface KlerosSubmission {
  id: string
  status: KlerosItemStatus
  listAddress: string
}
// ...
export interface Question {
  id: string
  raw: string
  templateId: number
  title: string
  resolution: Date
  arbitratorAddress: string
  category: string
  outcomes: string[]
  isPendingArbitration: boolean
  arbitrationOccurred: boolean
  currentAnswerTimestamp: Maybe<BigNumber>
  currentAnswerBond: Maybe<BigNumber>
  currentAnswer: string
  answers?: {
    answer: string
    bondAggregate: BigNumber
  }[]
  bonds?: BondItem[]
}

export enum OutcomeTableValue {
  OutcomeProbability = 'Outcome & Current Odds',
  CurrentPrice = 'Current Price',
  Shares = 'My Position',
  Payout = 'Payout',
  Outcome = 'Outcome',
  Probability = 'Odds',
  Bonded = 'Bonded',
}

export enum PositionTableValue {
  YourPosition = 'Your Position',
  Shares = 'Shares',
  Payout = 'Payout',
  ProfitLoss = 'Profit/loss',
}

export interface Token {
  address: string
  decimals: number
  symbol: string
  image?: string
  volume?: number
  balance?: string
  name?: string
  allowance?: string
}
// ...
export interface Arbitrator {
  address: string
  id: KnownArbitrator
  isSelectionEnabled: boolean
  name: string
  url: string
}
// ...
export interface MarketMakerData {
  address: string
  answerFinalizedTimestamp: Maybe<BigNumber>
  arbitrator: Arbitrator
  balances: BalanceItem[]
  creationTimestamp: string
  collateral: Token
  userInputCollateral: Token | null
  creator: string
  fee: BigNumber
  isConditionResolved: boolean
  isQuestionFinalized: boolean
  collateralVolume: BigNumber
  payouts: Maybe<Big[]>
  question: Question
  realitioAnswer: Maybe<BigNumber>
  totalEarnings: BigNumber
  totalPoolShares: BigNumber
  userEarnings: BigNumber
  userPoolShares: BigNumber
  klerosTCRregistered: boolean
  curatedByDxDao: boolean
  curatedByDxDaoOrKleros: boolean
  runningDailyVolumeByHour: BigNumber[]
  lastActiveDay: number
  scaledLiquidityParameter: number
  submissionIDs: KlerosSubmission[]
  oracle: string
  scalarLow: Maybe<BigNumber>
  scalarHigh: Maybe<BigNumber>
  outcomeTokenMarginalPrices: string[]
  outcomeTokenAmounts: string[]
}